twahifoundation/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
beautifulsoup4==4.9.1
boto==2.49.0
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.3.0
django-import-export==2.3.0
django-js-asset==1.2.2
django-libs==2.0.3
django-model-utils==4.0.0
django-mptt==0.10.0
django-notifications-hq==1.6.0
enum34==1.1.10
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
jdcal==1.4.1
jsonfield==3.1.0
lazy-object-proxy==1.4.3
ldif3==3.1.1
mailchimp-marketing==3.0.13
Markdown==3.2.2
MarkupPy==1.14
mccabe==0.6.1
microurl==0.1.1
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
pydot==1.4.1
pylint==2.5.3
pylint-django==2.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-card-me==0.9.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
South==1.0.2
sqlparse==0.3.1
surlex==0.2.0
swapper==1.1.2.post1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
unicode-slugify==0.1.3
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
